+++ src/backend/joanie/core/utils/payment_schedule.py
@@ -279,17 +279,21 @@ def get_paid_transactions(order):
 
 def get_transaction_references_to_refund(order) -> dict:
     """
-    Returns a dictionary containing transaction references as key and the amount of the installment
-    that are eligible to refund in an order's payment schedule.
+    Returns a dictionary containing transaction references as key that are eligible to be refunded
+    and the concerned installment as value.
     """
     to_refund_items = {}
+    used_installment_id = set()
     for transaction in get_paid_transactions(order):
         for installment in order.payment_schedule:
             if (
                 installment["state"] == enums.PAYMENT_STATE_PAID
                 and installment["amount"] == transaction.total
+                and transaction.reference not in to_refund_items
+                and installment["id"] not in used_installment_id
             ):
-                to_refund_items[transaction.reference] = installment["amount"]
+                to_refund_items[transaction.reference] = installment
+                used_installment_id.add(installment["id"])
     return to_refund_items